requirements.txt
Flask
gunicorn
pyrebase4
